+++ api/v1/routes.go
@@ -1,4 +1,4 @@
-package api
+package v1
 
 import (
 	"log"
