+++ api/v1/session_handler.go
@@ -1,4 +1,4 @@
-package api
+package v1
 
 import (
 	"net/http"
