+++ api/v1/tests/routes_test.go
@@ -1,4 +1,4 @@
-package api
+package v1
 
 import (
 	"net/http"
