+++ api/v1/tests/user_handler_test.go
@@ -1,4 +1,4 @@
-package api
+package v1
 
 import (
 	"bytes"
