+++ api/v1/user_handler.go
@@ -1,4 +1,4 @@
-package api
+package v1
 
 import (
 	"encoding/json"
